+++ cvat-ui/src/components/actions-menu/actions-menu.tsx
@@ -33,6 +33,7 @@ export enum Actions {
     EXPORT_TASK_DATASET = 'export_task_dataset',
     DELETE_TASK = 'delete_task',
     RUN_AUTO_ANNOTATION = 'run_auto_annotation',
+    MOVE_TASK_TO_PROJECT = 'move_task_to_project',
     OPEN_BUG_TRACKER = 'open_bug_tracker',
 }
 
@@ -128,6 +129,7 @@ export default function ActionsMenuComponent(props: Props): JSX.Element {
                 Automatic annotation
             </Menu.Item>
             <hr />
+            <Menu.Item key={Actions.MOVE_TASK_TO_PROJECT}>Move to project</Menu.Item>
             <Menu.Item key={Actions.DELETE_TASK}>Delete</Menu.Item>
         </Menu>
     );