+++ cvat-ui/src/reducers/interfaces.ts
@@ -74,6 +74,10 @@ export interface TasksState {
     fetching: boolean;
     updating: boolean;
     hideEmpty: boolean;
+    moveTask: {
+        modalVisible: boolean;
+        taskId: number | null;
+    };
     gettingQuery: TasksQuery;
     count: number;
     current: Task[];
@@ -245,6 +249,7 @@ export interface NotificationsState {
             exporting: null | ErrorState;
             deleting: null | ErrorState;
             creating: null | ErrorState;
+            moving: null | ErrorState;
         };
         formats: {
             fetching: null | ErrorState;
@@ -310,6 +315,7 @@ export interface NotificationsState {
     messages: {
         tasks: {
             loadingDone: string;
+            movingDone: string;
         };
         models: {
             inferenceDone: string;