+++ cvat-ui/src/reducers/tasks-reducer.ts
@@ -1,4 +1,4 @@
-// Copyright (C) 2020 Intel Corporation
+// Copyright (C) 2020-2021 Intel Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -14,6 +14,10 @@ const defaultState: TasksState = {
     fetching: false,
     updating: false,
     hideEmpty: false,
+    moveTask: {
+        modalVisible: false,
+        taskId: null,
+    },
     count: 0,
     current: [],
     gettingQuery: {
@@ -349,6 +353,16 @@ export default (state: TasksState = defaultState, action: AnyAction): TasksState
             return {
                 ...state,
                 hideEmpty: action.payload.hideEmpty,
+            };
+        }
+        case TasksActionTypes.SWITCH_MOVE_TASK_MODAL_VISIBLE: {
+            return {
+                ...state,
+                moveTask: {
+                    ...state.moveTask,
+                    modalVisible: action.payload.visible,
+                    taskId: action.payload.taskId,
+                },
             };
         }
         case BoundariesActionTypes.RESET_AFTER_ERROR: