+++ cvat/apps/engine/serializers.py
@@ -405,17 +405,76 @@ class TaskSerializer(WriteOnceMixin, serializers.ModelSerializer):
             instance.bug_tracker)
         instance.subset = validated_data.get('subset', instance.subset)
         labels = validated_data.get('label_set', [])
-        for label in labels:
-            LabelSerializer.update_instance(label, instance)
+        if instance.project_id is None:
+            for label in labels:
+                LabelSerializer.update_instance(label, instance)
+        validated_project_id = validated_data.get('project_id', None)
+        if validated_project_id is not None and validated_project_id != instance.project_id:
+            project = models.Project.objects.get(id=validated_data.get('project_id', None))
+            if instance.project_id is None:
+                for old_label in instance.label_set.all():
+                    try:
+                        new_label = project.label_set.filter(name=old_label.name).first()
+                    except ValueError:
+                        raise serializers.ValidationError(f'Target project does not have label with name "{old_label.name}"')
+                    old_label.attributespec_set.all().delete()
+                    for model in (models.LabeledTrack, models.LabeledShape, models.LabeledImage):
+                        model.objects.filter(job__segment__task=instance, label=old_label).update(
+                            label=new_label
+                        )
+                instance.label_set.all().delete()
+            else:
+                for old_label in instance.project.label_set.all():
+                    if new_label_for_name := list(filter(lambda x: x.get('id', None) == old_label.id, labels)):
+                        old_label.name = new_label_for_name[0].get('name', old_label.name)
+                    try:
+                        new_label = project.label_set.filter(name=old_label.name).first()
+                    except ValueError:
+                        raise serializers.ValidationError(f'Target project does not have label with name "{old_label.name}"')
+                    for (model, attr, attr_name) in (
+                        (models.LabeledTrack, models.LabeledTrackAttributeVal, 'track'),
+                        (models.LabeledShape, models.LabeledShapeAttributeVal, 'shape'),
+                        (models.LabeledImage, models.LabeledImageAttributeVal, 'image')
+                    ):
+                        attr.objects.filter(**{
+                            f'{attr_name}__job__segment__task': instance,
+                            f'{attr_name}__label': old_label
+                        }).delete()
+                        model.objects.filter(job__segment__task=instance, label=old_label).update(
+                            label=new_label
+                        )
+            instance.project = project
 
         instance.save()
         return instance
 
-    def validate_labels(self, value):
-        label_names = [label['name'] for label in value]
-        if len(label_names) != len(set(label_names)):
-            raise serializers.ValidationError('All label names must be unique for the task')
-        return value
+    def validate(self, attrs):
+        # When moving task labels can be mapped to one, but when not names must be unique
+        if 'project_id' in attrs.keys() and self.instance is not None:
+            project_id = attrs.get('project_id')
+            if project_id is not None and not models.Project.objects.filter(id=project_id).count():
+                raise serializers.ValidationError(f'Cannot find project with ID {project_id}')
+            # Check that all labels can be mapped
+            new_label_names = set()
+            old_labels = self.instance.project.label_set.all() if self.instance.project_id else self.instance.label_set.all()
+            for old_label in old_labels:
+                if len(new_labels := tuple(filter(lambda x: x.get('id') == old_label.id, attrs.get('label_set', [])))):
+                    new_label_names.add(new_labels[0].get('name', old_label.name))
+                else:
+                    new_label_names.add(old_label.name)
+            target_project = models.Project.objects.get(id=project_id)
+            target_project_label_names = set()
+            for label in target_project.label_set.all():
+                target_project_label_names.add(label.name)
+            if not new_label_names.issubset(target_project_label_names):
+                raise serializers.ValidationError('All task or project label names must be mapped to the target project')
+        else:
+            if 'label_set' in attrs.keys():
+                label_names = [label['name'] for label in attrs.get('label_set')]
+                if len(label_names) != len(set(label_names)):
+                    raise serializers.ValidationError('All label names must be unique for the task')
+
+        return attrs
 
 
 class ProjectSearchSerializer(serializers.ModelSerializer):